replace intmax_t and uintmax_t byt respectively  intptr_t and uintptr_t

+++ src/xtd.core.native.macos/src/xtd/native/macos/dns.mm
@@ -20,27 +20,27 @@ void dns::cleanup() {
   endhostent();
 }
 
-void dns::destroy(intmax_t host) {
+void dns::destroy(intptr_t host) {
   delete reinterpret_cast<hostent*>(host);
 }
 
-intmax_t dns::get_host_by_address(const string& host_address, int_least32_t host_address_type) {
+intptr_t dns::get_host_by_address(const string& host_address, int_least32_t host_address_type) {
   lock_guard<mutex> lock(dns_mutex);
   int_least64_t internet_address = 0;
   inet_pton(host_address_type, host_address.c_str(), &internet_address);
   hostent* host = gethostbyaddr(reinterpret_cast<char*>(&internet_address), host_address_type == ADDRESS_FAMILY_INTER_NETWORK ? 4 : 16, host_address_type);
   if (host == nullptr) return 0;
-  return reinterpret_cast<intmax_t>(new hostent(*host));
+  return reinterpret_cast<intptr_t>(new hostent(*host));
 }
 
-intmax_t dns::get_host_by_name(const string& host_name) {
+intptr_t dns::get_host_by_name(const string& host_name) {
   lock_guard<mutex> lock(dns_mutex);
   hostent* host = gethostbyname(host_name.c_str());
   if (host == nullptr) return 0;
-  return reinterpret_cast<intmax_t>(new hostent(*host));
+  return reinterpret_cast<intptr_t>(new hostent(*host));
 }
 
-vector<string> dns::get_aliases(intmax_t host) {
+vector<string> dns::get_aliases(intptr_t host) {
   vector<string> aliases;
   size_t index = 0;
   while (reinterpret_cast<hostent*>(host)->h_aliases[index] != nullptr)
@@ -48,7 +48,7 @@ vector<string> dns::get_aliases(intmax_t host) {
   return aliases;
 }
 
-vector<vector<uint_least8_t>> dns::get_addresses(intmax_t host) {
+vector<vector<uint_least8_t>> dns::get_addresses(intptr_t host) {
   size_t index = 0;
   vector<vector<uint_least8_t>> addresses;
   while (reinterpret_cast<hostent*>(host)->h_addr_list[index] != nullptr) {
@@ -58,7 +58,7 @@ vector<vector<uint_least8_t>> dns::get_addresses(intmax_t host) {
   return addresses;
 }
 
-string dns::get_host_name(intmax_t host) {
+string dns::get_host_name(intptr_t host) {
   return reinterpret_cast<hostent*>(host)->h_name;
 }
