+++ src/xtd.core.native.unix/src/xtd/native/unix/named_mutex.cpp
@@ -8,24 +8,24 @@
 
 using namespace xtd::native;
 
-intmax_t named_mutex::create(bool initially_owned, const std::string& name, bool& create_new) {
+intptr_t named_mutex::create(bool initially_owned, const std::string& name, bool& create_new) {
   sem_t* semaphore = nullptr;
   semaphore = sem_open(name.c_str(), O_CREAT | O_EXCL, S_IRUSR | S_IWUSR, 1);
   create_new = semaphore != SEM_FAILED;
   if (semaphore == SEM_FAILED) semaphore = sem_open(name.c_str(), O_CREAT, S_IRUSR | S_IWUSR, 1);
   if (semaphore != SEM_FAILED && initially_owned) {
     bool io_error = false;
-    wait(reinterpret_cast<intmax_t>(semaphore), -1, io_error);
+    wait(reinterpret_cast<intptr_t>(semaphore), -1, io_error);
   }
-  return semaphore != SEM_FAILED ? reinterpret_cast<intmax_t>(semaphore) : 0;
+  return semaphore != SEM_FAILED ? reinterpret_cast<intptr_t>(semaphore) : 0;
 }
 
-void named_mutex::destroy(intmax_t handle, const std::string& name) {
+void named_mutex::destroy(intptr_t handle, const std::string& name) {
   if (sem_close(reinterpret_cast<sem_t*>(handle)) == 0)
     sem_unlink(name.c_str());
 }
 
-bool named_mutex::signal(intmax_t handle, bool& io_error) {
+bool named_mutex::signal(intptr_t handle, bool& io_error) {
   io_error = false;
   if (sem_post(reinterpret_cast<sem_t*>(handle)) == -1) {
     if (errno == EINVAL) io_error = true;
@@ -34,7 +34,7 @@ bool named_mutex::signal(intmax_t handle, bool& io_error) {
   return true;
 }
 
-bool named_mutex::wait(intmax_t handle, int_least32_t milliseconds_timeout, bool& io_error) {
+bool named_mutex::wait(intptr_t handle, int_least32_t milliseconds_timeout, bool& io_error) {
   io_error = false;
   if (milliseconds_timeout == -1) {
     if (sem_wait(reinterpret_cast<sem_t*>(handle)) == -1) {
