+++ src/xtd.core.native/include/xtd/native/dns.h
@@ -50,33 +50,33 @@ namespace xtd {
       /// @brief Destroys host handle.
       /// @param host The host handle to destroy.
       /// @warning Internal use only
-      static void destroy(intmax_t host);
+      static void destroy(intptr_t host);
       /// @brief Gets host by host IP address.
       /// @param host_address The host IP address to resolve.
       /// @param host_address_type The  address family (see address_family_constants.h) to resolve.
       /// @return The host handle.
       /// @warning Internal use only
-      static intmax_t get_host_by_address(const std::string& host_address, int_least32_t host_address_type);
+      static intptr_t get_host_by_address(const std::string& host_address, int_least32_t host_address_type);
       /// @brief Gets host by host name.
       /// @param host_name The name to resolve.
       /// @return The host handle.
       /// @warning Internal use only
-      static intmax_t get_host_by_name(const std::string& host_name);
+      static intptr_t get_host_by_name(const std::string& host_name);
       /// @brief Gets aliases for a host.
       /// @param host The host handle to retrieve aliases.
       /// @return An alias array for yhe host.
       /// @warning Internal use only
-      static std::vector<std::string> get_aliases(intmax_t host);
+      static std::vector<std::string> get_aliases(intptr_t host);
       /// @brief Gets IP addresses for a host.
       /// @param host The host handle to retrieve IP addresses.
       /// @return An IP address array for yhe host.
       /// @warning Internal use only
-      static std::vector<std::vector<uint_least8_t>> get_addresses(intmax_t host);
+      static std::vector<std::vector<uint_least8_t>> get_addresses(intptr_t host);
       /// @brief Gets name for a host.
       /// @param host The host handle to retrieve name.
       /// @return A string that represent the host.
       /// @warning Internal use only
-      static std::string get_host_name(intmax_t host);
+      static std::string get_host_name(intptr_t host);
       /// @brief Gets host name for the current computer.
       /// @param host_name The string that will contain the host name.
       /// @return 0 if succeed; otherwise error.
